sss-idmap: add normalize and casefold options
If for id-mapping from a general source a string is used as input for
the murmurhash based offset calculation it might be important that
different UTF8 representations of the same string will lead to the same
ID (normalize) or that the case of the string does not matter for the ID
calculation (casefold).

By default normalization is enabled while casefolding is not.

:packaging: A new dependency to libutf8proc is added

Reviewed-by: Alexey Tikhonov <atikhono@redhat.com>
Reviewed-by: Iker Pedrosa <ipedrosa@redhat.com>