certmap: allow prefix in rule in sssd.conf
The current check for certificate mapping rules coming from sssd.conf is
too strict. Only rules which starts and ends with '(' and ')'
respectively are allowed. As a result new mapping templates with the
LDAPU1 prefix cannot be used.

With this patch prefixes are allowed as well. An existing integration
was was updated to cover this.

Resolves: https://github.com/SSSD/sssd/issues/7931

Reviewed-by: Alexey Tikhonov <atikhono@redhat.com>
Reviewed-by: Scott Poore <spoore@redhat.com>